doc: set in order plugins source
I think that use a variable to store the path of the sources files
is a good idea (let sourcepath = ...). With it, if I change the
plugin configuration directory, changing only one line will be
enough and all will works without much "effort", I guess.

It could be more useful use a loop and a collection (collection of
plugin's names) to "manage" these plugins. In other words, use some
logic here with Vim script. For now, how this looks make sense.