plugin/own/keybinds: update Gstatus map to open it on the right-side
It may be better to open Gstatus in a buffer on the right-side of
the screen because it's possible that you want to do a commit after
execute Gstatus, so you might press 'cc' and then the COMMIT_EDITMSG
will open on the top of the current fugitive buffer (Gstatus). You
will have a half of the scree to work on your commit. If you want to
return to the working buffer you would just press <C-w>h.