plugin/vim-go: add keybinds to test, run and build Golang
I wrote some vim-go keybinds to test, run, build Go's files and to
open Go's definitions and implementations.

I'm using the keyword's stats (at this moment I don't have its
percents):

    - top raw: neither more nor less used keys
    - midle raw: more used keys
    - bottom raw: less used keys

When I'm mapping a key or secuence of them, firts at all, I try to
figure out if that keybind will be often used or not, then in based
of it I select a raw to operate. But somethimes I prefer that the
keybinds make sense with its purpose. I'm still thinking about a
better combinations of keys to this purpose (test, run, build...).