plugin/own/keybinds: use <expr> else statement to change between errors
Instead using <C-m> and <C-n> to change between errors on the quickfix
window, use the else statement of the <expr>, that means, when vim is
not on diff mode, use <C-j> and <C-k> to change between errors. When
vim is on diff mode, use <C-j> and <C-k> to change between differences.