README: what's not latex.vim and what's its place
I've put recently latex.vim to this repository so I wanted to add its
explenation: what is not latex.vim (it's not a plugin) and what is
its place (it is in ./plugin/own/).

I wrote latex.vim and it only has two lines of code, it uses two tools:
pdflatex to compile from .tex to .pdf and mupdf to open and view PDFs.