init: change default opts of signcolumn and updatetime variables
signcolumn=yes in order to display, always, the signcolumn. If I don't use
this, when there is at least one file change, the signcolumn is added.  I
prefer always having that column there to avoid the animation (add and remove
the signcolumn).

the updatetime is set to a low value to stay up to date the signs of the
gitgutter plugin (when a add, remove or change some line of code).