plugin/own/keybinds: delete keybinds to change between buffers
For some reason, when I'm editing a commit message (like at this
moment), I press <C-h> (by mistake) and the commit is inmediatly
saved without I've finished to write it. This mistake (my mistake)
is awful, so I've decided to use "only" <Space> # to change between
buffers in order to avoid it.