init: enable filetype plugin and set a default colorcolumn value
I want to use a specific colorcolumn when I'm working with Go files
and I want to use other specific colorcolumn when I'm working with
tex files. In order to this, I've enabled filetype plugin and created
a ftplugin directory. With that it's possible to use differents
colorcolumn with differents file types.

By default, I'm using the colorcolum=80