plugin/own/keyinds: change <Leader> to <Space> to use fzf
I will use <Space> when the keybind is often used. Otherwise, when the
keybind is not often used, I'll use <Leader>. That's why I've changed
almost all the keybinds to search with fzf from <Leader> to <Space>.

    nnoremap <silent> <space>' :GCheckout<CR>
    nnoremap <silent> <leader>' :GCheckoutTag<CR>

I'm trying to make sense (also I know I'm not using mnemonics). The
last two keybinds are used to do a search between branches and tags,
they two have such the same purpose so I've used the same sign (') to
call one of them. The only change is that one use <Space> and the other
one use <Leader> before the sing (').